requirements.txt
pyperclip
Flask
flask-cors
pywebview
google-genai
pydantic
python-dotenv
keyboard
pyinstaller